Extract data tibbles from a REDCapTidieR supertibble and bind them to an environment

Take a supertibble generated with import_redcap() and bind its data tibbles (i.e. the tibbles in the redcap_data column) to an environment. The default is the global environment.

bind_tibbles(supertbl, environment = global_env(), tbls = NULL)

Arguments

supertbl

A supertibble generated by import_redcap(). Required.

environment

The environment to bind the tibbles to. Default is rlang::global_env().

tbls

A vector of the redcap_form_names of the data tibbles to bind to the environment. Default is NULL which binds all data tibbles.

Value

This function returns nothing as it's used solely for its side effect of modifying an environment.

Examples

# Create an empty environment
my_env <- new.env()

ls(my_env)
#> character(0)

# Mock up a supertibble
supertbl <- tibble::tribble(
  ~redcap_form_name,    ~redcap_data,   ~structure,
  "super_hero_powers",   list(),         "repeating",
  "heroes_information",  list(),         "nonrepeating"
)

bind_tibbles(supertbl, my_env)

ls(my_env)
#> [1] "heroes_information" "super_hero_powers"