Extract data tibbles from a REDCapTidieR supertibble into a list

Take a supertibble generated with import_redcap() and return a named list of data tibbles.

Usage

extract_tibbles(supertbl, tbls = everything())

Arguments

supertbl

A supertibble generated by import_redcap(). Required.

tbls

A vector of form_names or a tidyselect helper. Default is dplyr::everything().

Value

A named list of tibbles

Details

This function makes it easy to extract a multiple instrument's data from a REDCapTidieR supertibble into a named list. Specifying instruments using tidyselect helper functions such as dplyr::starts_with() or dplyr::ends_with() is supported.

Examples

# Mock up a supertibble
sample_data <- tibble::tribble(
   ~redcap_form_name,    ~redcap_data, ~structure,
   "super_hero_powers",   list(),       "repeating",
   "heroes_information",  list(),       "nonrepeating"
)

# Extract all data tibbles
extract_tibbles(sample_data)
#> $super_hero_powers
#> list()
#> 
#> $heroes_information
#> list()
#> 

# Only extract data tibbles starting with "heroes"
extract_tibbles(sample_data, starts_with("heroes"))
#> $heroes_information
#> list()
#>