Format REDCap variable labels

Use these functions with the format_labels argument of make_labelled() to define how variable labels should be formatted before being applied to the data columns of redcap_data. These functions are helpful to create pretty variable labels from REDCap field labels.

fmt_strip_whitespace() removes extra white space inside and at the start and end of a string. It is a thin wrapper of stringr::str_trim() and stringr::str_squish().

fmt_strip_trailing_colon() removes a colon character at the end of a string.

fmt_strip_trailing_punct() removes punctuation at the end of a string.

fmt_strip_html() removes html tags from a string.

fmt_strip_field_embedding() removes text between curly braces {} which REDCap uses for special "field embedding" logic. Note that read_redcap() removes html tags and field embedding logic from field labels in the metadata by default.

Usage

fmt_strip_whitespace(x)

fmt_strip_trailing_colon(x)

fmt_strip_trailing_punct(x)

fmt_strip_html(x)

fmt_strip_field_embedding(x)

Arguments

x

a character vector

Value

a modified character vector

Examples

fmt_strip_whitespace("Poorly Spaced   Label ")
#> [1] "Poorly Spaced Label"

fmt_strip_trailing_colon("Label:")
#> [1] "Label"

fmt_strip_trailing_punct("Label-")
#> [1] "Label"

fmt_strip_html("<b>Bold Label</b>")
#> [1] "Bold Label"

fmt_strip_field_embedding("Label{another_field}")
#> [1] "Label"

superheroes_supertbl
#> # A tibble: 2 × 9
#>   redcap_for…¹ redca…² redcap…³ redcap…⁴ struc…⁵ data_…⁶ data_…⁷ data_…⁸ data_…⁹
#>   <chr>        <chr>   <list>   <list>   <chr>     <int>   <int> <lbstr> <formt>
#> 1 heroes_info… Heroes… <tibble> <tibble> nonrep…     734      12 123.24… 0.23%  
#> 2 super_hero_… Super … <tibble> <tibble> repeat…    5966       4 180.21… 0%     
#> # … with abbreviated variable names ¹​redcap_form_name, ²​redcap_form_label,
#> #   ³​redcap_data, ⁴​redcap_metadata, ⁵​structure, ⁶​data_rows, ⁷​data_cols,
#> #   ⁸​data_size, ⁹​data_na_pct

make_labelled(superheroes_supertbl, format_labels = fmt_strip_trailing_colon)
#> # A tibble: 2 × 9
#>   redcap_for…¹ redca…² redcap…³ redcap…⁴ struc…⁵ data_…⁶ data_…⁷ data_…⁸ data_…⁹
#>   <chr>        <chr>   <list>   <list>   <chr>     <int>   <int> <lbstr> <formt>
#> 1 heroes_info… Heroes… <tibble> <tibble> nonrep…     734      12 123.24… 0.23%  
#> 2 super_hero_… Super … <tibble> <tibble> repeat…    5966       4 180.21… 0%     
#> # … with abbreviated variable names ¹​redcap_form_name, ²​redcap_form_label,
#> #   ³​redcap_data, ⁴​redcap_metadata, ⁵​structure, ⁶​data_rows, ⁷​data_cols,
#> #   ⁸​data_size, ⁹​data_na_pct