Determine fields included in REDCapR::redcap_read_oneshot output that should be dropped from results of read_redcap_tidy

Determine fields included in REDCapR::redcap_read_oneshot output that should be dropped from results of read_redcap_tidy

get_fields_to_drop(db_metadata, form)

Arguments

db_metadata

metadata tibble created by REDCapR::redcap_metadata_read

form

the name of the form containing identifiers

Value

A character vector of extra field names that can be used to filter the results of REDCapR::redcap_read_oneshot

Details

This function applies rules to determine which fields are included in the results of REDCapR::redcap_read_oneshot because the user didn't request the form containing identifiers