Extract REDCap Databases to Tidy Tibbles

Automatically detect REDCap database structure for the following elements:

Repeat Instruments

Longitudinal Events

Longitudinal Arms

Output a tibble with list elements containing tidy dataframes. Ideal for combination of tables with join operations primary and composite keys.

read_redcap_tidy(
  redcap_uri,
  token,
  raw_or_label = "label",
  suppress_messages = TRUE
)

Arguments

redcap_uri

The URI (uniform resource identifier) of the REDCap project. Required.

token

The user-specific string that serves as the password for a project. Required.

raw_or_label

A string (either 'raw' or 'label') that specifies whether to export the raw coded values or the labels for the options of multiple choice fields. Default is 'label'.

suppress_messages

Optionally show or suppress messages. Default TRUE.