Extract longitudinal REDCap databases into tidy tibbles

Helper function internal to read_redcap_tidy responsible for extraction and final processing of a tidy tibble to the user from a longitudinal REDCap database.

clean_redcap_long(db_data_long, db_metadata_long, linked_arms)

Arguments

db_data_long

The longitudinal REDCap database output defined by REDCapR::redcap_read_oneshot()$data

db_metadata_long

The longitudinal REDCap metadata output defined by REDCapR::redcap_metadata_read()$data

linked_arms

Output of link_arms, linking forms to REDCap events/arms

Value

Returns a tibble with list elements containing tidy dataframes. Users can access dataframes under the redcap_data column with reference to form_name and structure column details.